Finished Folder Refactoring
