--- .vs/VSWorkspaceState.json
@@ -1,8 +1,0 @@
-{
-  "ExpandedNodes": [
-    "",
-    "\\Controllers"
-  ],
-  "SelectedNode": "\\appsettings.Development.json",
-  "PreviewInSolutionExplorer": false
-}
