+++ .vscode/tasks.json
@@ -7,7 +7,7 @@
       "type": "process",
       "args": [
         "build",
-        "${workspaceFolder}/Shop.csproj",
+        "${workspaceFolder}/API/API.csproj",
         "/property:GenerateFullPaths=true",
         "/consoleloggerparameters:NoSummary"
       ],
@@ -19,7 +19,7 @@
       "type": "process",
       "args": [
         "publish",
-        "${workspaceFolder}/Shop.csproj",
+        "${workspaceFolder}/API/API.csproj",
         "/property:GenerateFullPaths=true",
         "/consoleloggerparameters:NoSummary"
       ],
@@ -33,7 +33,7 @@
         "watch",
         "run",
         "--project",
-        "${workspaceFolder}/Shop.csproj"
+        "${workspaceFolder}/API/API.csproj"
       ],
       "problemMatcher": "$msCompile"
     }
