+++ API/AuthenticationUtil.cs
