--- API/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace API;
-public class Class1
-{
-
-}
