+++ API/WebServiceExtensions.cs
@@ -1,3 +1,6 @@
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
+
 public static class WebServiceExtensions
 {
   public static void ConfigureIISIntegration(this IServiceCollection services) =>
