+++ Application/Application.csproj
@@ -1,12 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-  
-  <AssemblyName>Shop.Application</AssemblyName>
-  <Version>0.0.1</Version>
+
   </PropertyGroup>
 
 </Project>
