--- Application/Contracts/IServiceManager.cs
@@ -1,8 +1,0 @@
-namespace Application.Services.Contracts
-{
-
-  interface IServiceManager
-  {
-
-  }
-}
