--- Application/Services/ServiceManager.cs
@@ -1,9 +1,0 @@
-using Application.Services.Contracts;
-
-namespace Application.Services
-{
-  public class ServiceManager : IServiceManager
-  {
-
-  }
-}
