+++ Domain/Domain.csproj
@@ -1,12 +1,9 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-
-    <AssemblyName>Shop.Domain</AssemblyName>
-  <Version>0.0.1</Version>
   </PropertyGroup>
 
   <ItemGroup>
