--- Shop.csproj
@@ -1,50 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-
-    <AssemblyName>Shop</AssemblyName>
-    <Version>0.0.1</Version>
-
-  </PropertyGroup>
-
-  <ItemGroup>
-
-  
-      <Content Remove="Application\**" />
-      <Compile Remove="Application\**" />
-
-      <Content Remove="Domain\**" />
-      <Compile Remove="Domain\**" />
-
-      <Content Remove="Infrastructure\**" />
-      <Compile Remove="Infrastructure\**" />
-
-      <Content Remove="Contracts\**" />
-      <Compile Remove="Contracts\**" />
-  
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer.NetTopologySuite" Version="7.0.2" />
-  
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
-    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="6.0.13" />
-    <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="6.26.0" />
-    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.26.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="6.0.13" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.2" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.2">
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      <PrivateAssets>all</PrivateAssets>
-    </PackageReference>
-
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="Domain\Domain.csproj" />
-    <ProjectReference Include="Application\Application.csproj" />
-    <ProjectReference Include="Contracts\Contracts.csproj" />
-      </ItemGroup>
-
-</Project>
